+++ packages/plugin-packages/mashroom-background-jobs/src/js/errors/InvalidCronExpressionError.ts
@@ -1,0 +1,7 @@
+
+export default class InvalidCronExpressionError extends Error {
+    constructor(message?: string) {
+        super(message);
+        this.name = 'InvalidCronExpressionError';
+    }
+}