+++ packages/plugin-packages/mashroom-background-jobs/src/js/webapp/mashroom-bootstrap-webapp.ts
@@ -1,0 +1,13 @@
+
+import webapp from './webapp';
+
+import type {
+    MashroomWebAppPluginBootstrapFunction
+} from '@mashroom/mashroom/type-definitions';
+
+const bootstrap: MashroomWebAppPluginBootstrapFunction = async (pluginName, pluginConfig) => {
+
+    return webapp;
+};
+
+export default bootstrap;