+++ packages/test/test-server1/test-plugins/background-job/bootstrap.js
@@ -1,0 +1,14 @@
+
+
+const backgroundJob = (pluginContext) => {
+    const logger = pluginContext.loggerFactory('test.backgroundJob');
+
+    logger.info(`===== Executing dummy background job  =====`);
+};
+
+const bootstrap = () => {
+    return backgroundJob;
+};
+
+module.exports = bootstrap;
+