+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 
 ## [unreleased]
 
+ * Admin UI (under /mashroom) improved
+ * Added a new plugin type *admin-ui-integration* to register an arbitrary *web-app* or *static* plugin as panel in the Admin UI
+   (will be integrated via iframe)
+ * Remote App Registry Kubernetes: **BREAKING CHANGE**: Since it uses ow the new background job scheduler to scan the k8s cluster,
+   you also need to add *mashroom-background-jobs* to your dependencies, and the configuration slightly changed.
+   Checkout the README in *mashroom-portal-remote-app-registry-k8s*.
+ * Remote App Registry: **BREAKING CHANGE**: Since it uses ow the new background job scheduler to scan for remote apps,
+   you also need to add *mashroom-background-jobs* to your dependencies, and the configuration slightly changed.
+   Checkout the README in *mashroom-portal-remote-app-registry*.
+ * Added a background job plugin (*mashroom-background-jobs*) that allows it to schedule arbitrary jobs based on cron expressions
+ * HTTP Proxy: **BREAKING CHANGE**: The bootstrap for interceptor plugins must return the interceptor directly now instead of a Promise
  * HTTP Proxy: Added a second proxy implementation based on based on [node-http-proxy](https://github.com/request/request),
    since [request](https://github.com/http-party/node-http-proxy) is deprecated. It can be enabled like this in the config:
    ```json